requirements.txt
selenium
requests
python-dotenv
matplotlib
pandas
webdriver-manager
